% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcp-functions.R
\name{bcp_phenology_smooth}
\alias{bcp_phenology_smooth}
\title{Run Bayesian Changepoint Analysis using a single site and year based on loess smoothing}
\usage{
bcp_phenology_smooth(
  db.path,
  speciesID,
  templateID,
  locationID,
  amml,
  ensemble.type,
  span,
  ai.data,
  ai.variable,
  eval.data,
  keep.all.transition.days,
  year,
  start.date,
  end.date,
  lat,
  long,
  start.rise,
  end.rise,
  tz.recorder,
  tz.local,
  w0 = 0.2,
  p0 = 0.2,
  burnin = 500,
  mcmc = 1000,
  n.reps = 100
)
}
\arguments{
\item{db.path}{tbd}

\item{speciesID}{tbd}

\item{templateID}{tbd}

\item{locationID}{Site name of monitoring location for which to perform analysis}

\item{amml}{tbd}

\item{span}{LOESS smooth span to use (link to relevant details)}

\item{ai.data}{File path to acoustic index csv}

\item{eval.data}{File path to eval days csv that includes dates to exclude from analysis}

\item{keep.all.transition.days}{Logical flag for whether to keep ALL transition days (TRUE) or only keep the range (starting and ending) of transition days (FALSE)}

\item{year}{YYYY year for which to perform analysis}

\item{lat}{Latitude in decimal degrees, necessary to compute sunrise times}

\item{long}{Latitude in decimal degrees, necessary to computer sunrise times}

\item{start.rise}{Character object indicating how long, in hours, before sunrise to begin data subset (necessary if processing multiple years of data with varied audio recording schedules). See Details.}

\item{end.rise}{Character object indicating how long, in hours, after sunrise to end data subset (necessary if processing multiple years of data with varied audio recording schedules). See Details}

\item{tz.recorder}{Olsen names timezone used by the audio recorder during data collection. For example, you may have collected data using a Wildlife Acoustics SM4, and instead of setting a local time, UTC/GMT likely may have been used. Note that 'UTC' and 'GMT' are synonymous and both acceptable for this function argument. This argument accounts for the fact that recordings may have been taken in UTC. The tz.local argument then allows us to convert the times to local times that will make sense for analysis.}

\item{tz.local}{Olsen names timezone for local time at the monitoring location (e.g., 'America/Anchorage').}

\item{n.reps}{Number of times to repeat the BCP analysis to account for stochasticity}
}
\value{
Returns a list object with full results, peak days, and transition days. Give a detailed explanation of each item.
}
\description{
This function uses the bcp package to perform a Bayesian Changepoint Analysis for a desired site and year, based on loess smoothing of acoustic indices or species count data
}
\details{
This function was developed by the National Park Service Natural Sounds and Night Skies Division to perform a Bayesian Changepoint Analysis (BCP) in bioacoustics phenology projects for a selected year, acoustic monitoring site, and acoustic index metric.

To account for stochasticity, the BCP is performed n.reps times (default n.reps = 100).
Transition days are defined as the earliest and latest season date for which the average posterior probability falls above 0.5
Peak days are defined as the most frequently occurring(??) days with the highest posterior mean over n.reps (IS THIS HOW I SHOULD DEFINE IT?)

A note on timezones. SM4s do not account for daylight saving changes. The tz.local (local timezone) allows us to account for changes between Standard and Daylight time.

Because most bcp_phenology projects are going to be looking at sunrise-based audio sampling, the start.rise and end.rise arguments account for the possibility of varied sampling audio recording schemes through the years. E.g. for GLBA, start.rise = 1.5 and end.rise = 0.5 indicates that data subsetted beginning 1.5 hours before sunrise and ending 0.5 hours after sunrise.

Audio sampling that did not use a sunrise-based sampling scheme errs on the side of adding an additional ai timestep chunk on each end of the subset.

\itemize{
\item{Cool list of stuff}
\item{More stuff}
\item{and more!}
}
}
\examples{
\dontrun{

# examples here

}

}
\seealso{
\code{\link{NVSPL_To_AI}}, \code{\link{SongMeter_To_NVSPL}}
}
