% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_conf_threshold.R
\name{birdnet_conf_threshold}
\alias{birdnet_conf_threshold}
\title{(BETA) Generate options for BirdNET confidence thresholds}
\usage{
birdnet_conf_threshold(
  data,
  probabilities = c(0.85, 0.9, 0.95),
  min.conf,
  plot = TRUE,
  plot.folder
)
}
\arguments{
\item{data}{Data.table or data.frame of detection data with verifications.
Columns must include \code{common_name} (character), \code{confidence} (numeric) and a
\code{verify} column that indicates 1 for true positive and 0 for false positive.
The data type of the \code{verify} column does not matter as long as it is
composed of 1 and 0.}

\item{probabilities}{Vector of probability cutoffs to investigate.
Default = \code{c(0.85, 0.90, 0.95)}.}

\item{min.conf}{Input the minimum BirdNET confidence value that you
originally used for your data in BirdNET Analyzer. If left blank, the
function attempts to find a \code{min_conf} column from your BirdNET results in \code{data}.}

\item{plot}{Logical for whether to produce and display plot(s) in R.
Default = \code{TRUE}. Note: if you have many species, see \code{plot.folder} argument
for an option to save plots to a folder instead of viewing them in R.}

\item{plot.folder}{(Optional) If used, will save plots to a folder of your
choice for more organized viewing. Typically a better option than the \code{plot}
argument if you have many species.}
}
\value{
List with two data.tables/data.frames:

\strong{logistic_regression_results}: Table of confidence score thresholds to
use for different probability cutoffs.

\itemize{
\item{\strong{common_name}: Species common name.}
\item{\strong{pr_true}: Probability that a detection is a true positive.}
\item{\strong{confidence}: BirdNET confidence value to achieve the associated
probability of true positive.}
}

\strong{machine_learning_performance_results}: Table of confidence score
thresholds to use based on maximizing for F1, precision, or recall.

\itemize{
\item{\strong{common_name}: Species common name.}
\item{\strong{maximization_metric}: Machine learning performance evaluation
metric.}
\item{\strong{confidence}: BirdNET confidence value needed to maximize this
performance evaluation metric.}
\item{\strong{metric_score}: Machine learning performance evaluation score
for this metric (ranges from 0-1).}
}

Optionally also returns a set of three plots for each species:

\itemize{
\item{\strong{1:} Plot showing logistic regression BirdNET score thresholds
at the cutoffs given by the user in the \code{probabilities} argument.}
\item{\strong{2:} Plot illustrating machine learning performance metric scores
at different BirdNET confidence thresholds.}
\item{\strong{3:} Histograms showing BirdNET detection confidence score
distributions for verified true positives (1) and false positives (0).}
}
}
\description{
Generate options for BirdNET confidence thresholds based on
logistic regression a la \href{https://link.springer.com/article/10.1007/s10336-024-02144-5}{Wood and Kahl (2024)}
and/or standard machine learning performance evaluation metrics (i.e., choose
the lowest BirdNET confidence threshold that maximizes the precision and F1
score, or the highest confidence threshold that maximizes the recall score).
}
\details{
This function was developed by the National Park Service Natural Sounds and
Night Skies Division to assist with choosing BirdNET confidence thresholds
either based on logistic regression or based on machine learning performance
evaluation metrics (F1, precision, recall). Note that there may be multiple
BirdNET confidence thresholds that maximize the F1, precision, or recall. This
function returns the lowest possible BirdNET confidence threshold you can choose
that will still maximize the precision and the F1 score. It also returns the
highest possible BirdNET confidence threshold you can use while still maximizing
recall.
}
\examples{
\dontrun{

# Your input data.table/data.frame might have a variety of formats,
# but at minimum, needs columns 'common_name', 'confidence', and a 'verify'
# column with 1 for true positives and 0 for false positives.
# Depending on the verification scheme you used, you may need to do
# some minor reformatting first, as we illustrate in the example:

data(exampleVerified)
vers <- exampleVerified

# Modify the verification column to become 0s and 1s
vers[verify == 'n', verify := 0]
vers[verify == 'y', verify := 1]

# Generate confidence thresholds without generating plots
birdnet_conf_threshold(
   data = vers,
   probabilities = c(0.85, 0.90, 0.95, 0.99),
   plot = FALSE
)

# Generate confidence thresholds and plots
birdnet_conf_threshold(
   data = vers,
   probabilities = c(0.85, 0.90, 0.95, 0.99),
   plot = TRUE
)

# Generate confidence thresholds; save plots to folder instead of plotting in R
dir.create('example-conf-threshold-plots')
birdnet_conf_threshold(
   data = vers,
   probabilities = c(0.75, 0.85, 0.90, 0.95, 0.99),
   plot = FALSE,
   plot.folder = 'example-conf-threshold-plots'
)

# Delete all temporary example files when finished
unlink(x = 'example-conf-threshold-plots', recursive = TRUE)

}
}
\seealso{
\code{\link{birdnet_verify}}, \code{\link{birdnet_format}}
}
