% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_format.R
\name{birdnet_format}
\alias{birdnet_format}
\title{Format BirdNET outputs}
\usage{
birdnet_format(results.directory, timezone)
}
\arguments{
\item{results.directory}{Path to directory where raw BirdNET result files
have been stored. This directory should not contain anything aside from
BirdNET results. All
results in a single folder should either be .csv or .txt files from \code{\link{birdnet_analyzer}},
or .csv files generated by running BirdNET directly from the command line using
rtype = "r" (v1.5.1) or rtype = "csv" (v2+).}

\item{timezone}{Timezone setting used in the audio recorder (e.g, "GMT").
This argument allows you to specify the timezone shown by the wave filename.
If recordings were taken in local time at your study site, specify an
\href{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones#List}{Olson-names-formatted character timezone}
for the location (e.g., "America/Los_Angeles"). If recordings were taken in
GMT, you can put either "GMT" or "UTC" (both are acceptable in R for downstream
date-time formatting). This argument is critical to foster clarity in data
analysis through the years and across monitoring locations, because some
projects may vary across time and space as to whether the standard operating
procedure specifies recordings in GMT vs. local time.}
}
\value{
Saves a new formatted file of BirdNET results with filename inclusion
"\verb{_formatted_}" and suffix "csv" (or possibly "txt") depending on input.

Output contains the following columns:

\itemize{
\item{\strong{filepath}: Filepath for the processed audio file.}
\item{\strong{start}: Start time of detection in seconds.}
\item{\strong{end}: End time of detection in seconds.}
\item{\strong{scientific_name}: Species scientific name.}
\item{\strong{common_name}: Species common name.}
\item{\strong{confidence}: BirdNET's confidence level in this detection ranging from 0 (least confident) to 1 (most confident).}
\item{\strong{lat}: Latitude input used (if present in the data).}
\item{\strong{long}: Longitude input used (if present in the data).}
\item{\strong{week}: Week of year (if present in the data).}
\item{\strong{overlap}: Overlap input used (if present in the data).}
\item{\strong{sensitivity}: Sensitivity input used (if present in the data).}
\item{\strong{min_conf}: Minimum confidence input used (if present in the data).}
\item{\strong{species_list}: Species list used (if present in the data).}
\item{\strong{model}: BirdNET model used (if present in the data).}
\item{\strong{recordingID}: Recording identifier for the file, e.g., SITE_YYYYMMDD_HHMMSS.wav.}
\item{\strong{verify}: A column into which verifications may be populated. When initially created, will be 'NA'.}
\item{\strong{timezone}: Timezone setting used in the audio recorder.}
}
}
\description{
Reformats BirdNET outputs with a "recordingID" column for easier
data manipulation, a "verify" column to support manual verification of
detection results, a "timezone" column to clarify the timezone setting
used by the audio recorder, and R-friendly column headings.
}
\details{
This function was developed by the National Park Service Natural Sounds and
Night Skies Division to process audio data using BirdNET. It was developed
for use on Windows. CSV encodings may vary based on platform and result in errors.
}
\examples{

# Create a BirdNET results directory for this example
dir.create('example-results-directory')

# Write examples of raw BirdNET outputs to example results directory
data(exampleBirdNET1)
write.table(
  x = exampleBirdNET1,
  file = 'example-results-directory/Rivendell_20210623_113602.BirdNET.results.r.csv',
  row.names = FALSE, quote = FALSE, sep = ','
)

data(exampleBirdNET2)
write.table(
  x = exampleBirdNET2,
  file = 'example-results-directory/Rivendell_20210623_114602.BirdNET.results.r.csv',
  row.names = FALSE, quote = FALSE, sep = ','
)

# Run birdnet_format:
birdnet_format(
  results.directory = 'example-results-directory',
  timezone = 'GMT'
)

# Delete all temporary example files when finished
unlink(x = 'example-results-directory', recursive = TRUE)


}
\seealso{
\code{\link{birdnet_analyzer}}, \code{\link{birdnet_verify}}
}
