% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_plot_detections.R
\name{birdnet_plot_detections}
\alias{birdnet_plot_detections}
\title{Plot BirdNET detections}
\usage{
birdnet_plot_detections(
  data,
  audio.directory,
  title,
  frq.lim = c(0, 12),
  new.window = TRUE,
  spec.col = monitoR::gray.3(),
  box = TRUE,
  box.lwd = 1,
  box.col = "black"
)
}
\arguments{
\item{data}{Data.table or data.frame of subsetted detections that a user would like to plot \strong{for a single species}. This allows the user precise control over which detections to plot.}

\item{audio.directory}{tbd}

\item{title}{Optional title describing which detections are being plotted (e.g., "Confidence > 0.5", "True Positives", "Alarm Calls").}

\item{frq.lim}{Default = c(0, 12). Optional two-element numeric vector specifying frequency limits to the plotted spectrograms, in kHz.}

\item{new.window}{Default = TRUE. Logical value for whether to use \code{dev.new} to produce new plot windows.}

\item{spec.col}{Default = gray.3(). The colors used to plot verification spectrograms. Spectrogram colors are adjustable, and users may create their own gradients for display. A few spectrogram color options are provided via the R package monitoR, including gray.1(), gray.2(()), gray.3(), rainbow.1(), and topo.1(), all of which are based on existing R colors.}

\item{box}{Default = TRUE. Logical for whether to draw a box around each detection.}

\item{box.lwd}{Default = 1. Integer value for box line thickness.}

\item{box.col}{Default = 'black'. Box color.}
}
\value{
At least one plot, and up to N plots based on the number of "verify" categories contained in input data; one of all verified target signals for this template, and the other of all verified false alarms for this template. \code{birdnet_plot_detections} also returns a composite key of each plotted detection, which could be used to backtrack into the CSVs and check or change labels that appear to be incorrect... (not sure if I want to bother including this though).
}
\description{
tbd
}
\details{
This function was developed by the National Park Service Natural Sounds and Night Skies Division to process audio data produced by BirdNET.
}
\examples{
\dontrun{

# tbd
}

}
\seealso{
\code{\link{birdnet_run}}, \code{\link{birdnet_format_csv}}
}
