% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_review_verify.R
\name{birdnet_review_verify}
\alias{birdnet_review_verify}
\title{Save verifications from BirdNET GUI segment review to formatted CSVs}
\usage{
birdnet_review_verify(segments.directory, results.directory)
}
\arguments{
\item{segments.directory}{Segments directory path}

\item{results.directory}{Path to CSV result directory}
}
\value{
Updates the 'verify' column of CSVs generated from \code{\link{birdnet_format}}
with 'Positive' or 'Negative' for segments verified by a user via the Review tab of BirdNET
Analyzer GUI V2.
}
\description{
If a user has used the BirdNET Analyzer GUI Segments
tab to generate segments, and has used the GUI Review tab to provide verifications,
this function allows reviewed segment data to be saved with the labels
'Positive' and/or 'Negative' in the 'verify' column of formatted CSVs created
from \code{\link{birdnet_format}}. \strong{Use with caution: if verifications already exist in the 'verify'
column of the CSV, note that they may be overwritten by this function, or may
produce a mix of inconsistent labels.}
}
\details{
This function was developed by the National Park Service Natural Sounds and
Night Skies Division to process audio data using BirdNET.

Function has only been tested for BirdNET Analyzer GUI version 2+.
Unexpected behavior may result if attempting to combine BirdNET Analyzer V2 GUI
segment outputs with BirdNET Analyzer V1 GUI segment outputs.
}
\examples{

# Create a BirdNET results directory for this example
# Write examples of formatted BirdNET outputs to example results directory
dir.create('example-results-directory')
data(exampleFormatted1)
write.table(
   x = exampleFormatted1,
   file = 'example-results-directory/Rivendell_20210623_113602.BirdNET_formatted_results.csv',
   row.names = FALSE, quote = FALSE, sep = ',')
data(exampleFormatted2)
write.table(
  x = exampleFormatted2,
  file = 'example-results-directory/Rivendell_20210623_114602.BirdNET_formatted_results.csv',
  row.names = FALSE, quote = FALSE, sep = ',')

# Create directories with reviewed segments for this example
dir.create("example-segments-directory/Swainson's Thrush/Negative",
           recursive = TRUE)
dir.create("example-segments-directory/Swainson's Thrush/Positive",
           recursive = TRUE)
data(exampleSegments)
fi.names <- paste0("example-segments-directory/Swainson's Thrush/",
              c("Negative/0.224_67_Rivendell_20210623_113602_411.0s_414.0s.wav",
                "Negative/0.343_35_Rivendell_20210623_113602_588.0s_591.0s.wav",
                "Positive/0.666_7_Rivendell_20210623_113602_129.0s_132.0s.wav",
                "Positive/0.996_6_Rivendell_20210623_114602_345.0s_348.0s.wav"))
for (i in 1:length(exampleSegments)) {
tuneR::writeWave(
   object = exampleSegments[[i]],
   filename = fi.names[i]
  )
}

# Add reviewed segment verifications to "verify" column of formatted results
birdnet_review_verify(
  segments.directory = 'example-segments-directory',
  results.directory = 'example-results-directory'
)

# Delete all temporary example files when finished
unlink(x = 'example-results-directory', recursive = TRUE)
unlink(x = 'example-segments-directory', recursive = TRUE)

}
\seealso{
\code{\link{birdnet_format}}, \code{\link{birdnet_verify}}
}
