% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_run.R
\name{birdnet_run}
\alias{birdnet_run}
\title{Run BirdNET from RStudio}
\usage{
birdnet_run(
  audio.directory,
  results.directory,
  birdnet.directory,
  birdnet.script,
  lat,
  lon,
  ovlp = 0,
  sens = 1,
  min.conf = 0.1,
  custom.list = NULL
)
}
\arguments{
\item{audio.directory}{Absolute path to top-level input directory path to audio files to be processed. Files are expected to have the naming convention SITEID_YYYYMMDD_HHMMSS.wav.}

\item{results.directory}{Absolute path to directory where BirdNET results should be stored. Logical flag to specify whether audio files are housed in 'Data' subdirectories}

\item{birdnet.directory}{Absolute path to directory where BirdNET is installed on your machine.}

\item{birdnet.script}{Name of Python file that will be sourced to run BirdNET (e.g., 'analyze.py')}

\item{lat}{Recording location latitude. Set -1 to ignore.}

\item{lon}{Recording location latitude. Set -1 to ignore.}

\item{ovlp}{Overlap in seconds between extracted spectrograms. Values from 0.0 to 2.9. Default = 0.0.}

\item{sens}{Detection sensitivity; higher values result in higher sensitivity. Values from 0.5 to 1.5. Default = 1.0.}

\item{min.conf}{Minimum confidence threshold. Values from 0.01 to 0.99. Default = 0.1.}

\item{custom.list}{Not incorporated yet! From BirdNET:  Path to text file containing a list of species. Not used if not provided.}

\item{timezone}{Specify timezone setting used in the audio recorder (e.g, 'GMT'). If recordings were taken in local time at your study site, specify an \href{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones#List}{Olson-names-formatted character timezone} for the location (e.g., 'America/Los_Angeles'). This is extremely important to foster clarity in data analysis through the years, as some projects have varied year to year in whether recordings were taken in GMT vs. local time.}
}
\value{
Saves an unformatted CSV of results for each wave file in results.directory. Files have prefix "BirdNET_". If there is an issue with any audio files (e.g., file corrupted), problematic files that were not processed will be recorded in a file named 'BirdNET_Problem-Files_DATE.csv'.
}
\description{
This function uses the reticulate package to run Python from RStudio in order to process files through BirdNET. It assumes that all files in a folder come from the same site, and that the audio files are wave format and follow a SITEID_YYYYMMDD_HHMMSS.wav naming convention. To use this function, users must first successfully install \href{https://github.com/kahst/BirdNET-Lite}{BirdNET-Lite}, set up a conda environment, and make necessary modifications to the BirdNET analyze.py file. See \href{https://cbalantic.github.io/Install-BirdNET-Windows-RStudio/}{here} for one way to accomplish these steps (note that this does not serve as official guidance). Please input absolute paths for directory arguments. This is necessary due to the way RStudio is communicating with the underlying Python code.
}
\details{
This function was developed by the National Park Service Natural Sounds and Night Skies Division to act as a wrapper to process audio data using BirdNET.
}
\examples{
\dontrun{


Sys.setenv(RETICULATE_PYTHON = "C:/Your/Python/Path/Here/python.exe")
library(reticulate)
use_condaenv(condaenv = "pybirdnet", required = TRUE)

# can't really do an reproducible example here since I don't know what their install situation is
# birdnet_run()

}

}
\seealso{
\code{\link{birdnet_format_csv}}, \code{\link{birdnet_verify}}
}
