% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_verify.R
\name{birdnet_verify}
\alias{birdnet_verify}
\title{Verify BirdNET detections}
\usage{
birdnet_verify(
  data,
  audio.directory,
  results.directory,
  overwrite = FALSE,
  play = FALSE,
  buffer = 0,
  f.lim = c(0, 12),
  spec.col = monitoR::gray.3()
)
}
\arguments{
\item{data}{Data.table or data.frame of subsetted detections that a user would like to verify. This allows the user precise control over which detections to verify.}

\item{audio.directory}{tbd}

\item{results.directory}{tbd}

\item{overwrite}{tbd}

\item{play}{tbd}

\item{buffer}{tbd}

\item{f.lim}{tbd}

\item{spec.col}{tbd}
}
\value{
Saves a formatted CSV of results with the following columns, now updated to include any verifications by the user.

#' \itemize{
\item{\strong{recordingID}: Site name.}
\item{\strong{start.s}: tbd.}
\item{\strong{end.s}: tbd}
\item{\strong{scientific.name}: tbd.}
\item{\strong{common.name}: tbd.}
\item{\strong{confidence}: ....}
\item{\strong{verify}: ....}
\item{\strong{timezone}: ...}
}
}
\description{
tbd
}
\details{
This function was developed by the National Park Service Natural Sounds and Night Skies Division to process audio data using BirdNET.
}
\examples{
\dontrun{

# tbd
}

}
\seealso{
\code{\link{birdnet_run}}, \code{\link{birdnet_format_csv}}
}
