% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_verify_segments.R
\name{birdnet_verify_segments}
\alias{birdnet_verify_segments}
\title{(BETA) Verify BirdNET segment detections}
\usage{
birdnet_verify_segments(
  verification.library,
  segments.directory,
  results.directory,
  frq.lim = c(0, 12),
  spec.col = monitoR::gray.3()
)
}
\arguments{
\item{verification.library}{Character vector specifying which verification options should be shown to the user. Allows finer control to fit user's needs: user may specify whether a detection is a song, a call, a certain song or call type of interest, false alarm, unsure, or whatever the user needs. This enables maximum flexibility for the user, but also requires some thoughtfulness so that verification options remain consistent. BirdNET provides only species-level confidence and does not classify to song or call types. Thus, depending on the underlying research question and a verifier's expertise and familiarity with a focal species, a user may find themselves in a situation where songs are easily verified, but calls are not. Verification library provides the flexibility to accommodate varying questions and levels of expertise, but must be thought through by the user beforehand.}

\item{segments.directory}{Directory path for segments.}

\item{results.directory}{Directory path where you would like the output csv to be placed.}

\item{frq.lim}{Optional two-element numeric vector specifying frequency limits, in kHz, to apply to the plotted spectrograms. Default = c(0, 12).}

\item{spec.col}{The colors used to plot verification spectrograms. Default = gray.3(). Spectrogram colors are adjustable, and users may create their own gradients for display. A few spectrogram color options are provided via the R package monitoR, including gray.1(), gray.2(), gray.3(), rainbow.1(), and topo.1(), all of which are based on existing R colors.}
}
\value{
Returns a CSV of user-input verifications.
}
\description{
(BETA) Interactive function that produces spectrograms and wave clips enabling a user to verify BirdNET "segments".
}
\details{
This function was developed by the National Park Service Natural Sounds and
Night Skies Division to process and explore audio data using BirdNET.

Spectrograms show a segment detected by BirdNET. Title of spectrogram indicates
the segment audio file name.
}
\examples{
\dontrun{

}

}
\seealso{
\code{\link{birdnet_analyzer}}
}
