% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-helpfiles.R
\docType{data}
\name{exampleAI}
\alias{exampleAI}
\title{exampleAI}
\format{
An object of class \code{data.frame} containing 60 columns.
}
\usage{
data(exampleAI)
}
\description{
Sample data.frame to demonstrate \code{\link{plot_audio_summary}}
}
\details{
The exampleAI object is used to demonstrate \code{\link{plot_audio_summary}}
}
\examples{

# Load the dataset
data(exampleAI)

# View structure
str(exampleAI, max.level = 1)

}
\seealso{
\code{\link{plot_audio_summary}}, \code{\link{nvspl_to_ai}}
}
\keyword{datasets}
