% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-helpfiles.R
\docType{data}
\name{exampleAudio}
\alias{exampleAudio}
\title{exampleAudio}
\format{
An object of class \code{Wave} containing 6 slots.
}
\usage{
data(exampleAudio)
}
\description{
Sample recording (audio) data stored as a \code{Wave} file.
}
\details{
The exampleAudio object can be used to experiment with functions that use audio data, such as \code{\link{SongMeter_To_NVSPL}}
}
\examples{
# Load the dataset
data(exampleAudio)

# View structure
str(exampleAudio, max.level = 1)

}
\keyword{datasets}
