% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-helpfiles.R
\docType{data}
\name{exampleAudio1}
\alias{exampleAudio1}
\title{exampleAudio1}
\format{
An object of class \code{Wave} containing 6 slots.
}
\usage{
data(exampleAudio1)
}
\description{
Sample recording (audio) data stored as a \code{Wave} file.
}
\details{
The exampleAudio1 object can be used to experiment with functions that use audio data, such as \code{\link{wave_to_nsvpl}}
}
\examples{
# Load the dataset
data(exampleAudio1)

# View structure
str(exampleAudio1, max.level = 1)

}
\keyword{datasets}
