% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-helpfiles.R
\docType{data}
\name{exampleAudio2}
\alias{exampleAudio2}
\title{exampleAudio2}
\format{
An object of class \code{Wave} containing 6 slots.
}
\usage{
data(exampleAudio2)
}
\description{
Sample recording (audio) data stored as a \code{Wave} file.
}
\details{
The exampleAudio2 object can be used to experiment with functions that use audio data, such as \code{\link{wave_to_nvspl}}
}
\examples{
# Load the dataset
data(exampleAudio2)

# View structure
str(exampleAudio2, max.level = 1)

}
\keyword{datasets}
