% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-helpfiles.R
\docType{data}
\name{exampleBarchartData}
\alias{exampleBarchartData}
\title{exampleBarchartData}
\format{
An object of class \code{data.frame} and \code{data.table} containing 17 columns.
}
\usage{
data(exampleBarchartData)
}
\description{
Sample data.table / data.frame to demonstrate \code{\link{birdnet_barchart}}
}
\details{
The exampleBarchartData object is used to demonstrate \code{\link{birdnet_barchart}}
}
\examples{

# Load the dataset
data(exampleBarchartData)

# View structure
str(exampleBarchartData, max.level = 1)

}
\keyword{datasets}
