% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-helpfiles.R
\docType{data}
\name{exampleBirdNET1}
\alias{exampleBirdNET1}
\title{exampleBirdNET1}
\format{
An object of class \code{data.frame} containing 14 columns.
}
\usage{
data(exampleBirdNET1)
}
\description{
Sample of raw BirdNET output.
}
\details{
The exampleBirdNET1 object can be used to experiment with birdnet_* functions, such as \code{\link{birdnet_format}}
}
\examples{
# Load the dataset
data(exampleBirdNET1)

# View structure
str(exampleBirdNET1, max.level = 1)

# Create a BirdNET results directory for this example
dir.create('example-results-directory')

# Write examples of raw BirdNET txt outputs to example results directory to
# mimic BirdNET output format
write.table(x = exampleBirdNET1,
          file = 'example-results-directory/BirdNET_Rivendell_20210623_113602.txt',
          row.names = FALSE, quote = FALSE, sep = ',')

# Delete all temporary example files when finished
unlink(x = 'example-results-directory', recursive = TRUE)

}
\keyword{datasets}
