% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-helpfiles.R
\docType{data}
\name{exampleDatesSampled}
\alias{exampleDatesSampled}
\title{exampleDatesSampled}
\format{
An object of class \code{Date} of length 430.
}
\usage{
data(exampleDatesSampled)
}
\description{
Example data for \code{\link{birdnet_heatmap}} and \code{\link{birdnet_heatmap_time}}
}
\details{
The exampleDatesSampled object can be used to experiment with
birdnet_heatmap* functions.
}
\examples{
# Load the dataset
data(exampleDatesSampled)

# View structure
str(exampleDatesSampled, max.level = 1)


}
\keyword{datasets}
