% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-helpfiles.R
\docType{data}
\name{exampleFormatted1}
\alias{exampleFormatted1}
\title{exampleFormatted1}
\format{
An object of class \code{data.frame} containing 17 columns.
}
\usage{
data(exampleFormatted1)
}
\description{
Sample of formatted BirdNET output.
}
\details{
The exampleFormatted1 object can be used to experiment with birdnet_*
functions, such as \code{\link{birdnet_gather}}
}
\examples{
# Load the dataset
data(exampleFormatted1)

# View structure
str(exampleFormatted1, max.level = 1)

# Create a BirdNET results directory for this example
dir.create('example-results-directory')

# Write examples of formatted BirdNET outputs to example results directory
# to mimic output format of birdnet_format
write.table(
x = exampleFormatted1,
file = 'example-results-directory/Rivendell_20210623_113602.BirdNET_formatted_results.csv',
            row.names = FALSE, quote = FALSE, sep = ','
)

# Delete all temporary example files when finished
unlink(x = 'example-results-directory', recursive = TRUE)

}
\keyword{datasets}
