% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-helpfiles.R
\docType{data}
\name{exampleHeatmapData}
\alias{exampleHeatmapData}
\title{exampleHeatmapData}
\format{
An object of class \code{data.table} and \code{data.frame} containing 14 columns.
}
\usage{
data(exampleHeatmapData)
}
\description{
Example data for \code{\link{birdnet_heatmap}} and \code{\link{birdnet_heatmap_time}}
}
\details{
The exampleHeatmapData object can be used to experiment with birdnet_heatmap* functions.
}
\examples{
# Load the dataset
data(exampleHeatmapData)

# View structure
str(exampleHeatmapData, max.level = 1)


}
\keyword{datasets}
