% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-helpfiles.R
\docType{data}
\name{exampleNVSPL}
\alias{exampleNVSPL}
\title{exampleNVSPL}
\format{
An object of class \code{list} containing 9 data.frames.
}
\usage{
data(exampleNVSPL)
}
\description{
Sample NVSPL data stored as data.frames within a list object.
}
\details{
The exampleNVSPL object can be used to experiment with functions that use NVSPL data. See the example code given in \code{\link{NVSPL_To_AI}}.
}
\examples{

# Create an input directory for this example
dir.create('example-input-directory')

# Read in example NVSPL data
data(exampleNVSPL)

# View structure
str(exampleNVSPL, 1)

# Write example NVSPL data to example input directory to mimic NVSPL .txt format
for (i in 1:length(exampleNVSPL)) {
write.table(x = exampleNVSPL[[i]],
            file = paste0('example-input-directory/', names(exampleNVSPL)[i]),
            sep = ',',
            quote = FALSE)
}

# Delete all temporary example files when finished
unlink(x = 'example-input-directory', recursive = TRUE)

}
\keyword{datasets}
