% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-helpfiles.R
\docType{data}
\name{examplePlotData}
\alias{examplePlotData}
\title{examplePlotData}
\format{
An object of class \code{data.frame} and \code{data.table} containing 17 columns.
}
\usage{
data(examplePlotData)
}
\description{
Sample data.frame to demonstrate \code{\link{birdnet_plot}}
}
\details{
The examplePlotData object is used to demonstrate \code{\link{birdnet_plot}}
}
\examples{

# Load the dataset
data(examplePlotData)

# View structure
str(examplePlotData, max.level = 1)

}
\keyword{datasets}
