% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-helpfiles.R
\docType{data}
\name{exampleScores}
\alias{exampleScores}
\title{exampleScores}
\format{
An object of class \code{data.frame} and \code{data.table} containing 4 columns.
}
\usage{
data(exampleScores)
}
\description{
Sample scores data.frame.
}
\details{
The exampleScores object is used to demonstrate \code{\link{add_time_cols}}.
}
\examples{

# Read in example data
data(exampleScores)

# View structure
str(exampleScores, 1)

}
\keyword{datasets}
