% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-helpfiles.R
\docType{data}
\name{exampleSegments}
\alias{exampleSegments}
\title{exampleSegments}
\format{
An object of class \code{list} containing 4 \code{Wave} objects each with 6 slots.
}
\usage{
data(exampleSegments)
}
\description{
Sample audio segments stored in a list as \code{Wave} files.
}
\details{
The exampleSegments object can be used to experiment with functions
that use audio data, such as \code{\link{birdnet_review_verify}}.
}
\examples{
# Load the dataset
data(exampleSegments)

# View structure
str(exampleSegments, max.level = 1)

}
\keyword{datasets}
