% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-helpfiles.R
\docType{data}
\name{exampleSpeciesList}
\alias{exampleSpeciesList}
\title{exampleSpeciesList}
\format{
An object of class \code{data.table} and \code{data.frame} with 4 rows and 1 column.
}
\usage{
data(exampleSpeciesList)
}
\description{
Example data for \code{\link{birdnet_analyzer}}
}
\details{
The exampleSpeciesList object can be used to test the \code{\link{birdnet_analyzer}} function.
}
\examples{
# Load the dataset
data(exampleSpeciesList)

# View structure
str(exampleSpeciesList, max.level = 1)


}
\keyword{datasets}
