% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-helpfiles.R
\docType{data}
\name{exampleSpectroData}
\alias{exampleSpectroData}
\title{exampleSpectroData}
\format{
An object of class \code{data.frame} and \code{data.table} containing 17 columns.
}
\usage{
data(exampleSpectroData)
}
\description{
Sample data.frame to demonstrate \code{\link{birdnet_spectro}}
}
\details{
The exampleSpectroData object is used to demonstrate \code{\link{birdnet_spectro}}
}
\examples{

# Load the dataset
data(exampleSpectroData)

# View structure
str(exampleSpectroData, max.level = 1)

}
\keyword{datasets}
