% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-helpfiles.R
\docType{data}
\name{exampleVerified}
\alias{exampleVerified}
\title{exampleVerified}
\format{
An object of class \code{data.table} and \code{data.frame} with 828
rows and 17 variables.
}
\usage{
data(exampleVerified)
}
\description{
Example data for \code{\link{birdnet_conf_threshold}}
}
\details{
The exampleVerified object illustrates what a typical data.frame /
data.table looks like with a column of verification data from the \code{\link{birdnet_verify}}
function. It can also be used to test the \code{\link{birdnet_conf_threshold}}
function.
}
\examples{
# Load the dataset
data(exampleVerified)

# View structure
str(exampleVerified, max.level = 1)


}
\keyword{datasets}
