% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta.R
\name{Meta}
\alias{Meta}
\title{Internal function from PAMGuide code.}
\usage{
Meta(
  ...,
  atype = "TOL",
  plottype = "Both",
  envi = "Air",
  calib = 0,
  stype = "MF",
  Si = -159,
  Mh = -36,
  G = 0,
  vADC = 1.414,
  r = 50,
  N = Fs,
  winname = "Hann",
  lcut = Fs/N,
  hcut = Fs/2,
  timestring = "",
  outwrite = 1,
  welch = "",
  chunksize = "",
  linlog = "Log",
  filenms = filenms
)
}
\description{
Internal function from PAMGuide code.
}
\keyword{internal}
