% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvspl_to_ai.R
\name{nvspl_to_ai}
\alias{nvspl_to_ai}
\title{Convert NVSPLs data into acoustic indices}
\usage{
nvspl_to_ai(
  input.directory,
  output.directory,
  project,
  instrum = "SM4",
  start.at.beginning = TRUE,
  fminimum = "H1600",
  fmaximum = "H8000",
  fbinMax = 8,
  BKfminimum = "H31p5",
  BKfmaximum = "H1250",
  timestep = 10,
  startday = 1,
  plt = FALSE
)
}
\arguments{
\item{input.directory}{Top-level NVSPL directory from which to ingest data (often named "NVSPL").}

\item{output.directory}{Where to place output files (often named "ANALYSIS").}

\item{project}{File name for your project (e.g., 'GLBAPhenology2019').}

\item{instrum}{Audio recorder used. Default = 'SM4'.}

\item{start.at.beginning}{Logical flag for where computation should start. Default = TRUE.  If TRUE, computation starts at the beginning of recordings on a given day. TRUE is preferred in most cases for NSNSD bioacoustic analyses. When FALSE, indices are calculated only for data that begins at the top of the chosen timestep. For example, if timestep is 10 minutes, then each AI calculation will start at 0, 10, 20, 30, 40, 50 for each hour. If the full ten minute sample is not available, the timestep is skipped. start.at.beginning = FALSE may be desired if the user wishes to break results easily into hours.}

\item{fminimum}{Lower limit of frequency band of interest (NOTE: HXXXX represents the center frequency of the octave band, eg. H1600 = 1413-1778 Hz). Default = 'H1600'.}

\item{fmaximum}{Upper limit of frequency band of interest. Default = 'H8000'.}

\item{fbinMax}{Sets the frequency bins to look across for cluster analysis. Default = 8.}

\item{BKfminimum}{Lower limit of frequency bands for background noise (NOTE: if NVSPL calculated with PAMGuide, no data below H25). Default = 'H31p5'.}

\item{BKfmaximum}{Upper limit of frequency bands for background noise. Default = 'H1250'.}

\item{timestep}{Timestep of computation, in minutes (NOTE: to get daily values, summarizing the timestep values is preferred over running the code for an entire day). Default = 10.}

\item{startday}{Optional start date if many files or error.}

\item{plt}{Logical flag for whether to plot NVSPL files. NOTE: this will slow the function down considerably. Default = FALSE.}
}
\value{
Returns a CSV file of the acoustic indices to the output directory. For posterity, also saves a "params" file to the output.directory.
}
\description{
Convert NVSPL table data into acoustic indices.
}
\details{
CSV file contains the following 59 columns:

\itemize{
\item{\strong{Site}: Site name.}
\item{\strong{Date}: Date as YYYY_MM_DD.}
\item{\strong{Yr}: Year as YYYY.}
\item{\strong{Mo}: Month of year as integer.}
\item{\strong{Day}: Day of month.}
\item{\strong{Hr}: Hour of day.}
\item{\strong{Min}: Minute of hour}
\item{\strong{Sec}: Second of minute.}
\item{\strong{timestep}: In minutes, the time increment over which acoustic indices are calculated.}
\item{\strong{SampleLength_sec}: Sample length (i.e., timestep) in seconds.}
\item{\strong{ACIout}: ACI results for each time chunk on a given day - SPLs  (link to documentation of how this was computed, or describe here) }
\item{\strong{ACIoutI}: ACI results for each time chunk on a given day - intensity (unlog)}
\item{\strong{ACIoutN}: ACI results for each time chunk on a given day -  normalized }
\item{\strong{BKdB_low}: These are all background noise of some kind. Unclear on the details based on the code }
\item{\strong{BKdBA_low}: tbd. }
\item{\strong{BKdB_bird}: tbd. }
\item{\strong{BKdBA_bird}: tbd. }
\item{\strong{avgAMP}: tbd. }
\item{\strong{L10AMP}: tbd. }
\item{\strong{Hf}: tbd. }
\item{\strong{Ht}: tbd. }
\item{\strong{El}: tbd. }
\item{\strong{BioPh}: tbd. }
\item{\strong{AthPh}: tbd. }
\item{\strong{SoundScapeI}: tbd. }
\item{\strong{AA}: Acoustic activity, but unclear to me what each of these means based on the code. Regular AA is average time a noise is above background in each timestep. }
\item{\strong{AAc}: tbd. }
\item{\strong{AAdur}: tbd. }
\item{\strong{AAanth}: tbd. }
\item{\strong{AAcanth}: tbd. }
\item{\strong{AAduranth}: tbd. }
\item{\strong{Rough}: Roughness - change in acoustic energy from one second to the next, summarized over timestep. }
\item{\strong{ADI_step}: Acoustic diversity index for the timestep (based on vegan::diversity) }
\item{\strong{Eveness_step}: ineq::Gini }
\item{\strong{pk}: tbd. }
\item{\strong{pkd}: tbd. }
\item{\strong{pks}: tbd. }
\item{\strong{Hm}: tbd. }
\item{\strong{HvPres}: tbd. }
\item{\strong{HvSPL}: tbd. }
\item{\strong{unlist(dif_L10L90)}: tbd. What is this colname?}
\item{\strong{Mamp}:  maybe this is acoustic richness instead, according to the code (though at some point I concluded not based on plots) }
\item{\strong{NumCL}: tbd. }
\item{\strong{SP2}: this might be spectral persistence?? }
\item{\strong{CL1dur}: tbd. }
\item{\strong{CL1pk}: tbd. }
\item{\strong{CL1Leq}: tbd. }
\item{\strong{CL2dur}: tbd. }
\item{\strong{CL2pk}: tbd. }
\item{\strong{CL2Leq}: tbd. }
\item{\strong{CL3dur}: tbd. }
\item{\strong{CL3pk}: tbd. }
\item{\strong{CL3Leq}: tbd. }
\item{\strong{CL4dur}: tbd. }
\item{\strong{CL4pk}: tbd. }
\item{\strong{CL4Leq}: tbd. }
\item{\strong{vers}: PAMGUIDE version used for NVSPLs that gets carried over into here? Or is this AI versioning, documentation of which I have yet to locate? }
\item{\strong{AR}: Acoustic richness - i'm assuming that's what AR stands for? }
}
}
\examples{
\dontrun{

# Create an input and output directory for this example
dir.create('example-input-directory')
dir.create('example-output-directory')

# Read in example NVSPL data
data(exampleNVSPL)

# Write example NVSPL data to example input directory
for (i in 1:length(exampleNVSPL)) {
  write.table(
    x = exampleNVSPL[[i]],
    file = paste0('example-input-directory/', names(exampleNVSPL)[i]),
    sep = ',',
    quote = FALSE
  )
}

# Run nvspl_to_ai to generate acoustic indices csv for example NVSPL files
nvspl_to_ai(
  input.directory = 'example-input-directory',
  output.directory = 'example-output-directory',
  project = 'example-project'
)

# View Results
(ai.results <- read.csv(
  list.files(path = 'example-output-directory',
             pattern = '.csv', full.names = TRUE))
)

# Delete all temporary example files when finished
unlink(x = 'example-input-directory', recursive = TRUE)
unlink(x = 'example-output-directory', recursive = TRUE)

}

}
\seealso{
\code{\link{wave_to_nvspl}}
}
