% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAMGuide_Meta.R
\name{PAMGuide_Meta}
\alias{PAMGuide_Meta}
\title{Internal function from PAMGuide code.}
\usage{
PAMGuide_Meta(
  fullfile,
  ...,
  atype = "TOL",
  plottype = "Both",
  envi = "Air",
  calib = 0,
  ctype = "TS",
  Si = -159,
  Mh = -36,
  G = 0,
  vADC = 1.414,
  r = 50,
  N = Fs,
  winname = "Hann",
  lcut = Fs/N,
  hcut = Fs/2,
  timestring = "",
  outdir = dirname(fullfile),
  outwrite = 0,
  disppar = 1,
  welch = "",
  chunksize = "",
  linlog = "Log"
)
}
\description{
Internal function from PAMGuide code.
}
\keyword{internal}
