% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcp-functions.R
\name{phenology_smooth}
\alias{phenology_smooth}
\title{Use LOESS smoothing to track phenology.}
\usage{
phenology_smooth(
  db.path,
  speciesID,
  templateID,
  locationID,
  amml,
  ensemble.type,
  span,
  ai.data,
  ai.variable,
  eval.data,
  year,
  start.date,
  end.date,
  lat,
  long,
  start.rise,
  end.rise,
  tz.recorder,
  tz.local
)
}
\arguments{
\item{db.path}{tbd}

\item{speciesID}{tbd}

\item{templateID}{tbd}

\item{locationID}{Site name of monitoring location for which to perform analysis}

\item{amml}{tbd}

\item{span}{LOESS smooth span to use (link to relevant details)}

\item{ai.data}{File path to acoustic index csv}

\item{eval.data}{File path to eval days csv that includes dates to exclude from analysis}

\item{year}{YYYY year for which to perform analysis}

\item{lat}{Latitude in decimal degrees, necessary to compute sunrise times}

\item{long}{Latitude in decimal degrees, necessary to computer sunrise times}

\item{start.rise}{Character object indicating how long, in hours, before sunrise to begin data subset (necessary if processing multiple years of data with varied audio recording schedules). See Details.}

\item{end.rise}{Character object indicating how long, in hours, after sunrise to end data subset (necessary if processing multiple years of data with varied audio recording schedules). See Details}

\item{tz.recorder}{Olsen names timezone used by the audio recorder during data collection. For example, you may have collected data using a Wildlife Acoustics SM4, and instead of setting a local time, UTC/GMT likely may have been used. Note that 'UTC' and 'GMT' are synonymous and both acceptable for this function argument. This argument accounts for the fact that recordings may have been taken in UTC. The tz.local argument then allows us to convert the times to local times that will make sense for analysis.}

\item{tz.local}{Olsen names timezone for local time at the monitoring location (e.g., 'America/Anchorage').}
}
\value{
Returns a list object with stuff...
}
\description{
Create loess smoothing model to capture trends of acoustic indices or species count data. Simple version with no BCP analysis necessary; peak days can be identified directly from peak in loess curve; transition periods can be identified based on thoughtful and biologically meaningful definitions of transitions.
}
\details{
add deetzs

\itemize{
\item{Cool list of stuff}
\item{More stuff}
\item{and more!}
}
}
\examples{
\dontrun{

# examples here

}

}
\seealso{
\code{\link{NVSPL_To_AI}}, \code{\link{SongMeter_To_NVSPL}}
}
