% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Viewer.R
\name{Viewer}
\alias{Viewer}
\title{Internal function from PAMGuide code.}
\usage{
Viewer(..., plottype = "Both", fullfile = "", ifile = "", linlog = "Log")
}
\description{
Internal function from PAMGuide code.
}
\keyword{internal}
