% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_time_cols}
\alias{add_time_cols}
\title{Add R-friendly time-based columns to bioacoustics data}
\usage{
add_time_cols(dt, recording.id.col = "recordingID", tz.recorder, tz.local)
}
\arguments{
\item{dt}{Data.frame or data.table containing a recordingID column of format
LOCATIONID_YYYYMMDD_HHMMSS.wav}

\item{recording.id.col}{Column name that contains recordingID with format
'LOCATIONID_YYYYMMDD_HHMMSS.wav'. Default = 'recordingID'.}

\item{tz.recorder}{Olsen names timezone used by the audio recorder during data
collection. For example, you may have collected data using a Wildlife Acoustics
SM4, and may have used UTC/GMT instead of setting a local time. Note
that 'UTC' and 'GMT' are synonymous and both acceptable for this function
argument. This argument accounts for the fact that recordings may have been
taken in UTC. The tz.local argument then allows us to convert the times to
local times that will make sense for analysis.}

\item{tz.local}{Olsen names timezone for local time at the monitoring
location (e.g., 'America/Anchorage').}
}
\value{
A data.table with the same columns as the input, but now including
the following additional columns:
\itemize{
\item{\strong{dateTimeRecorder}: POSIXct-formatted date-time object used by
the audio recorder (typically will be the same as UTC or local time, depending
on recording configuration parameters used in the field).}
\item{\strong{dateTimeUTC}: POSIXct-formatted date-time object for the recordingID, in UTC time.}
\item{\strong{dateTimeLocal}: POSIXct-formatted date-time object for the recordingID in local time.}
\item{\strong{date}: POSIXct-formatted date-time object in UTC time.}
\item{\strong{year}: POSIXct-formatted date-time object in local time.}
\item{\strong{detectionTimeLocal}: POSIXct-formatted date-time object for the precise time of a detection, in local time.}
\item{\strong{locationID}: Character value for the locationID.}

}
}
\description{
Convenience function to parse recordingIDs into useful columns and
enable expedient summaries by date and local time. Input table must
have a parseable recordingID column of format LOCATIONID_YYYYMMDD_HHMMSS.wav
}
\details{
This function was developed by the National Park Service Natural Sounds and
Night Skies Division to support bioacoustics projects.
}
\examples{

# Ensure your data has an appropriate recordingID column and time columns
dat <- exampleHeatmapData
dat[ , recordingID := basename(filepath)]
dat <- add_time_cols(
 dt = dat,
 tz.recorder = 'America/Los_angeles',
 tz.local = 'America/Los_angeles'
)

}
