% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_barchart.R
\name{birdnet_barchart}
\alias{birdnet_barchart}
\title{Plot stacked barcharts of BirdNET detections}
\usage{
birdnet_barchart(
  data,
  julian.breaks,
  interactive = FALSE,
  focal.species,
  focal.colors
)
}
\arguments{
\item{data}{Data.table or data.frame of formatted BirdNET results that a user would like to plot. Generally, this data object may be preceded by a call to \code{\link{add_time_cols}}; all data should come from a single site and the object must contain columns named "locationID" (character), "recordingID" (character), and "dateTimeLocal" (POSIXct).}

\item{julian.breaks}{Optional numeric vector of julian date plotting breaks to use on the x-axis. If omitted, will be computed automatically.}

\item{interactive}{Default = FALSE for a static plot. If true, produces a plotly plot with interactive hover.}

\item{focal.species}{Optional character vector containing the common names of species to highlight. Any species contained in the data object that are not contained in focal.species will be plotted under one color as "Other".}

\item{focal.colors}{Optional character vector describing the colors that should be used for species named in the focal.species argument. Non-focal species ("Other") will be plotted in black.}
}
\value{
Stacked barchart of BirdNET results
}
\description{
Plot stacked barcharts of user-selected BirdNET results by date
}
\details{
This function was developed by the National Park Service Natural Sounds and Night Skies Division. It is intended to provide exploratory plotting for summarizing and visualizing BirdNET results.

Generally, interactive = FALSE should be used in conjunction with the focal.species argument. If focal.species argument is provided, a legend will also be plotted.

Use of interactive = TRUE is meant strictly for exploratory purposes. If focal.species argument is not used, no legend will be plotted. A typical use case is to omit focal.species when setting interactive = TRUE.
}
\examples{
\dontrun{

# Read in exampleBarchartData
data(exampleBarchartData)

# Generally, add_time_cols() may be called as part of preprocessing
# (if not, please ensure data object has columns that include locationID (character),
# recordingID (character), and dateTimeLocal (POSIXct))
dat <- add_time_cols(dt = exampleBarchartData,
                     tz.recorder = 'America/Los_angeles',
                     tz.local = 'America/Los_angeles')

# Produce an interactive plotly barchart with interactive = TRUE
birdnet_barchart(data = dat, interactive = TRUE)

# Produce a static ggplot barchat with interactive = FALSE,
# add focal.species with custom colors (any species in the data object
# that are not in focal.species will be plotted in black as "Other".)
birdnet_barchart(data = dat,
                 interactive = FALSE,
                 focal.species = c("Pacific Wren", "Swainson's Thrush", "Varied Thrush"),
                 focal.colors = c('#00BE67', '#C77CFF', '#c51b8a'))

}

}
\seealso{
\code{\link{birdnet_analyzer}}, \code{\link{birdnet_format}}, \code{\link{birdnet_verify}}
}
