% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_format_csv.R
\name{birdnet_format_csv}
\alias{birdnet_format_csv}
\title{Format BirdNET csv outputs}
\usage{
birdnet_format_csv(results.directory, timezone)
}
\arguments{
\item{results.directory}{Path to directory where raw BirdNET result CSVs have been stored}

\item{timezone}{Specify timezone setting used in the audio recorder (e.g, 'GMT'). If recordings were taken in local time at your study site, specify an \href{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones#List}{Olson-names-formatted character timezone} for the location (e.g., 'America/Los_Angeles'). This is extremely important to foster clarity in data analysis through the years, as some projects have varied year to year in whether recordings were taken in GMT vs. local time.}
}
\value{
Saves a formatted CSV of results with the following columns

#' \itemize{
\item{\strong{recordingID}: Site name.}
\item{\strong{start.s}: tbd.}
\item{\strong{end.s}: tbd}
\item{\strong{scientific.name}: tbd.}
\item{\strong{common.name}: tbd.}
\item{\strong{confidence}: ....}
\item{\strong{verify}: ....}
\item{\strong{timezone}: ...}
}
}
\description{
tbd
}
\details{
This function was developed by the National Park Service Natural Sounds and Night Skies Division to process audio data using BirdNET.
}
\examples{
\dontrun{

# tbd
}

}
\seealso{
\code{\link{birdnet_run}}, \code{\link{birdnet_verify}}
}
