% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_format_csv.R
\name{birdnet_format_csv}
\alias{birdnet_format_csv}
\title{Format BirdNET CSV outputs}
\usage{
birdnet_format_csv(results.directory, timezone)
}
\arguments{
\item{results.directory}{Path to directory where raw BirdNET result CSVs have been stored}

\item{timezone}{Timezone setting used in the audio recorder (e.g, "GMT"). This argument allows you to specify the timezone shown by the wave filename. If recordings were taken in local time at your study site, specify an \href{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones#List}{Olson-names-formatted character timezone} for the location (e.g., "America/Los_Angeles"). If recordings were taken in GMT, you can put either "GMT" or "UTC" (both are acceptable in R for downstream date-time formatting). This argument is critical to foster clarity in data analysis through the years and across monitoring locations, because some projects may vary across time and space as to whether the standard operating procedure specifies recordings in GMT vs. local time.}
}
\value{
Saves a new formatted CSV of BirdNET results with filename prefix "BirdNET_formatted_" and the following columns:

\itemize{
\item{\strong{recordingID}: Recording identifier for the file, as SITE_YYYYMMDD_HHMMSS.wav.}
\item{\strong{start.s}: Start time of detection in seconds.}
\item{\strong{end.s}: End time of detection in seconds.}
\item{\strong{scientific.name}: Species scientific name.}
\item{\strong{common.name}: Species common name.}
\item{\strong{confidence}: BirdNET's confidence level in this detection ranging from 0 (least confident) to 1 (most confident).}
\item{\strong{verify}: A column into which verifications may be populated. When initially created, will be 'NA'.}
\item{\strong{timezone}: Timezone setting used in the audio recorder.}
}
}
\description{
Reformats BirdNET CSV outputs into separated columns with R-friendly column names, a "recordingID" column for easier data manipulation, a "verify" column to support manual verification of detection results, and a "timezone" column to clarify the timezone setting used by the audio recorder.
}
\details{
This function was developed by the National Park Service Natural Sounds and Night Skies Division to process audio data using BirdNET. It was developed for use on Windows. CSV encodings may vary based on platform and result in errors.
}
\examples{
\dontrun{

# Create a BirdNET results directory for this example
dir.create('example-results-directory')

# Write examples of raw BirdNET CSV outputs to example results directory
data(exampleBirdNET1)
write.csv(x = exampleBirdNET1,
          file = 'example-results-directory/BirdNET_Rivendell_20210623_113602.csv',
          row.names = FALSE, )

data(exampleBirdNET2)
write.csv(x = exampleBirdNET2,
          file = 'example-results-directory/BirdNET_Rivendell_20210623_114602.csv',
          row.names = FALSE)

# Run birdnet_format_csv:
birdnet_format_csv(results.directory = 'example-results-directory',
                   timezone = 'GMT')

# Delete all temporary example files when finished
unlink(x = 'example-results-directory', recursive = TRUE)

}

}
\seealso{
\code{\link{birdnet_run}}, \code{\link{birdnet_verify}}
}
