% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_gather_results.R
\name{birdnet_gather_results}
\alias{birdnet_gather_results}
\title{Gather BirdNET results}
\usage{
birdnet_gather_results(results.directory)
}
\arguments{
\item{results.directory}{Path to directory where raw BirdNET result CSVs have been stored}
}
\value{
Returns a data.frame/data.table with the following columns:

\itemize{
\item{\strong{recordingID}: Site name.}
\item{\strong{start.s}: tbd.}
\item{\strong{end.s}: tbd}
\item{\strong{scientific.name}: tbd.}
\item{\strong{common.name}: tbd.}
\item{\strong{confidence}: ....}
\item{\strong{verify}: ....}
\item{\strong{timezone}: ...}
}
}
\description{
Gather all BirdNET CSV results from a desired folder into a user-friendly data.table / data.frame
}
\details{
This function was developed by the National Park Service Natural Sounds and Night Skies Division to process audio data using BirdNET.
}
\examples{
\dontrun{

# tbd
}

}
\seealso{
\code{\link{birdnet_run}}, \code{\link{birdnet_verify}}, \code{\link{birdnet_format_csv}}
}
