% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_species_counts.R
\name{birdnet_species_counts}
\alias{birdnet_species_counts}
\title{Provide count data of detection species over a selected time unit.}
\usage{
birdnet_species_counts(dt, unit)
}
\arguments{
\item{dt}{Path to directory where raw BirdNET result CSVs have been stored}

\item{unit}{See lubridate::round_date for options. E.g. second, 2 hours, etc.}
}
\value{
Saves a formatted CSV of results with the following columns:

\itemize{
\item{\strong{stuff}: tbd}
}
}
\description{
tbd
}
\details{
This function was developed by the National Park Service Natural Sounds and Night Skies Division to summarize audio data using BirdNET.
}
\examples{
\dontrun{

# tbd
}

}
\seealso{
\code{\link{birdnet_run}}, \code{\link{birdnet_verify}}, \code{\link{birdnet_format_csv}}
}
