% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_verify.R
\name{birdnet_verify}
\alias{birdnet_verify}
\title{Verify BirdNET detections}
\usage{
birdnet_verify(
  data,
  verification.library,
  audio.directory,
  results.directory,
  overwrite = FALSE,
  play = FALSE,
  buffer = 0,
  f.lim = c(0, 12),
  spec.col = monitoR::gray.3()
)
}
\arguments{
\item{data}{Data.table or data.frame of subsetted detections that a user would like to verify. This allows the user precise control over which detections to verify.}

\item{verification.library}{List specifying which verification options should be shown to the user. Allows finer control for user to specify whether a detection is a song, a call, a certain song or call type of interest, false alarm, unsure, or whatever the user needs. This allows maximum flexibility for the user, but also requires some thoughtfulness so that verification options remain consistent.}

\item{audio.directory}{tbd}

\item{results.directory}{Path to directory where raw BirdNET result CSVs have been stored}

\item{overwrite}{tbd}

\item{play}{tbd}

\item{buffer}{tbd}

\item{f.lim}{tbd}

\item{spec.col}{tbd}
}
\value{
Saves a formatted CSV of results with the following columns, now updated to include any verifications by the user.

#' \itemize{
\item{\strong{recordingID}: Site name.}
\item{\strong{start.s}: tbd.}
\item{\strong{end.s}: tbd}
\item{\strong{scientific.name}: tbd.}
\item{\strong{common.name}: tbd.}
\item{\strong{confidence}: ....}
\item{\strong{verify}: ....}
\item{\strong{timezone}: ...}
}
}
\description{
tbd
}
\details{
This function was developed by the National Park Service Natural Sounds and Night Skies Division to process audio data using BirdNET.

Spectrograms show 3-second segment detected by BirdNET. Title of spectrogram indicates the recordingID of the file name, the start and end times of the detection in seconds, the species detection, and the confidence level of the detection from 0 to 1.
}
\examples{
\dontrun{

# tbd
}

}
\seealso{
\code{\link{birdnet_run}}, \code{\link{birdnet_format_csv}}
}
