% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-helpfiles.R
\docType{data}
\name{exampleBirdNET2}
\alias{exampleBirdNET2}
\title{exampleBirdNET2}
\format{
An object of class \code{data.frame} containing 14 columns.
}
\usage{
data(exampleBirdNET2)
}
\description{
Sample of raw BirdNET output.
}
\details{
The exampleBirdNET2 object can be used to experiment with birdnet_* functions, such as \code{\link{birdnet_format}}
}
\examples{
# Load the dataset
data(exampleBirdNET2)

# View structure
str(exampleBirdNET2, max.level = 1)

# Create a BirdNET results directory for this example
dir.create('example-results-directory')

# Write examples of raw BirdNET txt outputs to example results directory to
# mimic BirdNET output format
write.table(x = exampleBirdNET2,
          file = 'example-results-directory/Rivendell_20210623_114602.BirdNET.results.csv',
          row.names = FALSE, quote = FALSE, sep = ',')

# Delete all temporary example files when finished
unlink(x = 'example-results-directory', recursive = TRUE)

}
\keyword{datasets}
