% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-helpfiles.R
\docType{data}
\name{exampleFormatted2}
\alias{exampleFormatted2}
\title{exampleFormatted2}
\format{
An object of class \code{data.frame} containing 17 columns.
}
\usage{
data(exampleFormatted2)
}
\description{
Sample of formatted BirdNET output.
}
\details{
The exampleFormatted2 object can be used to experiment with birdnet_* functions, such as \code{\link{birdnet_gather}}
}
\examples{
# Load the dataset
data(exampleFormatted2)

# View structure
str(exampleFormatted2, max.level = 1)

# Create a BirdNET results directory for this example
dir.create('example-results-directory')
# Write examples of formatted BirdNET outputs to example results directory
# to mimic output format of birdnet_format
write.table(x = exampleFormatted2,
            file = 'example-results-directory/BirdNET_formatted_Rivendell_20210623_114602.txt',
            row.names = FALSE, quote = FALSE, sep = ',')

# Delete all temporary example files when finished
unlink(x = 'example-results-directory', recursive = TRUE)

}
\keyword{datasets}
