% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVSPLtoAI-functions.R
\name{NVSPL_To_AI}
\alias{NVSPL_To_AI}
\title{Convert NVSPL to acoustic indices}
\usage{
NVSPL_To_AI(
  input.directory,
  output.directory,
  project,
  instrum = "SM4",
  start.at.beginning = FALSE,
  fminimum = "H1600",
  fmaximum = "H8000",
  fbinMax = 8,
  BKfminimum = "H31p5",
  BKfmaximum = "H1250",
  sunrise = 0,
  timestep = 10,
  startday = 1,
  plt = 0,
  savFile = 1
)
}
\arguments{
\item{input.directory}{Top-level NVSPL directory from which to ingest data (possibly "NVSPL")}

\item{output.directory}{Where to place output files, (possibly "ANALYSIS")}

\item{project}{File name for your project (e.g., 'GLBAPhenology2019')}

\item{instrum}{Audio recorder used (e.g., 'SM4')}

\item{start.at.beginning}{Logical flag for where computation should start. Default = FALSE. When FALSE, indices are calculated only for data that begins at the top of the chosen timestep.
For example, if your time step is 10 minutes, then each AI calculation will start at 0, 10, 20, 30, 40, 50 for each hour. If the full ten minute sample is not available, the timestep is skipped. This version is ideal if you want to break results easily into hours.
If TRUE, simply starts at the start of your recordings on a given day. TRUE is what is used for GLBA Phenology. ((CB: I THINK this means it doesn't throw away any sub-timestep chunks? Based on GLBA SOP))}

\item{fminimum}{Lower limit of frequency band of interest (NOTE: HXXXX represents the center frequency of the octave band, eg. H1600 = 1413-1778 Hz)}

\item{fmaximum}{Upper limit of frequency band of interest}

\item{fbinMax}{Sets the frequency bins to look across for cluster analysis (???)}

\item{BKfminimum}{Lower limit of frequency bands for background noise (NOTE: if NVSPL calculated with PAMGuide, no data below H25)}

\item{BKfmaximum}{Upper limit of frequency bands for background noise}

\item{sunrise}{Change to 1 if only want to calculate indices over sunrise (Need more explanation for what this is doing??)}

\item{timestep}{in minutes (NOTE: to get daily values, summarizing the timestep values is preferred over running the code for an entire day) - CB - so, number of minutes over which to calculate?}

\item{startday}{Optional start date if many files or error (CB: NOT A ROBUST WAY TO HANDLE ERRORS. ADD IN CONTINGENCIES FOR THIS AS I FIND THEM)}

\item{plt}{Option to plot NVSPL files (cb I think 1 = plot?) (NOTE: this will slow things way down!)}

\item{savFile}{(0 = false) Flag for whether to save file for each site with row for each timestep and AI values (? i.e., the CSV?)}
}
\value{
If savFile = 1, returns a CSV file of the acoustic indices to the output directory. For posterity, also saves a "params" file to the output.directory.
}
\description{
Convert NVSPLs (generated by PAMGuide) into acoustic indices. This version uses the code from "calculate_NVSPLtoAI_Generic.R" located under Synology / GLBA Phenology / June2020Upload
}
\details{
CSV file contains the following 59 columns:

\itemize{
\item{\strong{Site}: Site name.}
\item{\strong{Date}: Date as YYYY_MM_DD.}
\item{\strong{Yr}: Year as YYYY.}
\item{\strong{Mo}: Month of year as integer.}
\item{\strong{Day}: Day of month.}
\item{\strong{Hr}: Hour of day.}
\item{\strong{Min}: Minute of hour}
\item{\strong{Sec}: Second of minute.}
\item{\strong{timestep}: In minutes, the time increment over which acoustic indices are calculated.}
\item{\strong{SampleLength_sec}: Sample length (i.e., timestep) in seconds.}
\item{\strong{ACIout}: ACI results for each time chunk on a given day - SPLs  (link to documentation of how this was computed, or describe here) }
\item{\strong{ACIoutI}: ACI results for each time chunk on a given day - intensity (unlog)}
\item{\strong{ACIoutN}: ACI results for each time chunk on a given day -  normalized }
\item{\strong{BKdB_low}: These are all background noise of some kind. Unclear on the details based on the code }
\item{\strong{BKdBA_low}: stuff. }
\item{\strong{BKdB_bird}: stuff. }
\item{\strong{BKdBA_bird}: stuff. }
\item{\strong{avgAMP}: stuff. }
\item{\strong{L10AMP}: stuff. }
\item{\strong{Hf}: stuff. }
\item{\strong{Ht}: stuff. }
\item{\strong{El}: stuff. }
\item{\strong{BioPh}: stuff. }
\item{\strong{AthPh}: stuff. }
\item{\strong{SoundScapeI}: stuff. }
\item{\strong{AA}: Acoustic activity, but unclear to me what each of these means based on the code. Regular AA is average time a noise is above background in each timestep. }
\item{\strong{AAc}: stuff. }
\item{\strong{AAdur}: stuff. }
\item{\strong{AAanth}: stuff. }
\item{\strong{AAcanth}: stuff. }
\item{\strong{AAduranth}: stuff. }
\item{\strong{Rough}: Roughness - change in acoustic energy from one second to the next, summarized over timestep. }
\item{\strong{ADI_step}: Acoustic diversity index for the timestep (based on vegan::diversity) }
\item{\strong{Eveness_step}: ineq::Gini }
\item{\strong{pk}: stuff. }
\item{\strong{pkd}: stuff. }
\item{\strong{pks}: stuff. }
\item{\strong{Hm}: stuff. }
\item{\strong{HvPres}: stuff. }
\item{\strong{HvSPL}: stuff. }
\item{\strong{unlist(dif_L10L90)}: stuff. What is this colname?}
\item{\strong{Mamp}:  maybe this is acoustic richness instead, according to the code (though at some point I concluded not based on plots) }
\item{\strong{NumCL}: stuff. }
\item{\strong{SP2}: this might be spectral persistence?? }
\item{\strong{CL1dur}: stuff. }
\item{\strong{CL1pk}: stuff. }
\item{\strong{CL1Leq}: stuff. }
\item{\strong{CL2dur}: stuff. }
\item{\strong{CL2pk}: stuff. }
\item{\strong{CL2Leq}: stuff. }
\item{\strong{CL3dur}: stuff. }
\item{\strong{CL3pk}: stuff. }
\item{\strong{CL3Leq}: stuff. }
\item{\strong{CL4dur}: stuff. }
\item{\strong{CL4pk}: stuff. }
\item{\strong{CL4Leq}: stuff. }
\item{\strong{vers}: PAMGUIDE version used for NVSPLs that gets carried over into here? Or is this AI versioning, documentation of which I have yet to locate? }
\item{\strong{AR}: Acoustic richness - i'm assuming that's what AR stands for? }
}
}
\examples{
\dontrun{

# Create an input and output directory for this example
dir.create('example-input-directory')
dir.create('example-output-directory')

# Read in example NVSPL data
data(exampleNVSPL)

# Write example NVSPL data to example input directory
for (i in 1:length(exampleNVSPL)) {
write.table(x = exampleNVSPL[[i]],
            file = paste0('example-input-directory/', names(exampleNVSPL)[i]),
            sep = ',',
            quote = FALSE)
}

# Run NVSPL_To_AI to generate acoustic indices csv for example NVSPL files,
# setting start.at.beginning = FALSE
NVSPL_To_AI(input.directory = 'example-input-directory',
            output.directory = 'example-output-directory',
            project = 'example-project',
            start.at.beginning = FALSE)

# Run NVSPL_To_AI to generate acoustic indices csv for example NVSPL files,
# setting start.at.beginning = TRUE
NVSPL_To_AI(input.directory = 'example-input-directory',
            output.directory = 'example-output-directory',
            project = 'example-project',
            start.at.beginning = TRUE)

# Read in both files to compare
start.standard <- read.csv(list.files(path = './example-output-directory/',
                                     pattern = 'Index_Created',
                                     full.names = TRUE))
start.begin <- read.csv(list.files(path = './example-output-directory/',
                                     pattern = 'Begin_Created',
                                     full.names = TRUE))

# View a few rows of each file; note the Hr, Min, Sec differences between both options
start.standard[1:4, ]
start.begin[1:4, ]

# Delete all temporary example files when finished
unlink(x = 'example-input-directory', recursive = TRUE)
unlink(x = 'example-output-directory', recursive = TRUE)

}

}
\seealso{
\code{\link{Wave_To_NVSPL}}
}
