% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wave_To_NVSPL.R
\name{wave_to_nvspl}
\alias{wave_to_nvspl}
\title{Calibrate and convert wave files into NVSPL format}
\usage{
wave_to_nvspl(
  input.directory,
  data.directory = TRUE,
  test.file = FALSE,
  project,
  instrum = "SM4",
  filext = "_\%Y\%m\%d_\%H\%M\%S.wav",
  filpat = ".+\\\\d{8}_\\\\d{6}.wav",
  mhset = -35,
  Gset = 16,
  vADCset = 1,
  enviset = "Air",
  rescWat = 0,
  timezone,
  ...
)
}
\arguments{
\item{input.directory}{Top-level input directory path to audio files to be processed. e.g. E:/AUDIO. \strong{Audio files are expected to have the naming convention SITEID_YYYYMMDD_HHMMSS.wav. Use the filext argument for other file extension patterns (NOT TESTED BY CB).}}

\item{data.directory}{Logical flag to specify whether audio files are housed in 'Data' subdirectories}

\item{test.file}{Logical flag for whether to test a file. If TRUE, tests a single file and produces plots and diagnostic outputs. If FALSE, processes entire audio dataset indicated by input.directory.}

\item{project}{File name for your project (e.g., 'GLBAPhenology2019')}

\item{instrum}{Audio recorder used (e.g., 'SM4')}

\item{filext}{File extension pattern (e.g, '\emph{\%Y\%m\%d}\%H\%M\%S.wav'). If using split files, '\emph{0}\%Y\%m\%d_\%H\%M\%S_000.wav'.}

\item{filpat}{File pattern (can more info be given here? Or could we just constrain the expected file pattern so minimize unexpected user behavior?)}

\item{mhset}{Microphone sensitivity dBV/Pa (see for more info https://doimspp.sharepoint.com/sites/nsnsdallstaff/Shared\%20Documents/Science\%20and\%20Tech/Software/SongMeterToNVSPL/SongMeter4toNVSPL.mp4)}

\item{Gset}{Gain settings}

\item{vADCset}{Zero-peak   (more info? what does this mean)}

\item{enviset}{Use 'Air' or 'Wat' to indicate whether audio recordings occurred in a terrestrial or underwater environment, respectively.}

\item{rescWat}{Use 1 if you want to re-scale underwater values to be able to plot using AMT, 0 if not}

\item{timezone}{Specify timezone setting used in the audio recorder (e.g, 'GMT'). If recordings were taken in local time at your study site, specify an \href{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones#List}{Olson-names-formatted character timezone} for the location (e.g., 'America/Los_Angeles'). This is extremely important to foster clarity in data analysis through the years, as some projects have varied year to year in whether recordings were taken in GMT vs. local time.}

\item{...}{Other arguments to PAMGuide internal functions PAMGuide.R, PAMGuide_Meta.R, and Meta.R. See PAMGuide documentation for details.}
}
\value{
If test.file = TRUE, returns diagnostics. If test.file = FALSE, returns NVSPL txt files in NVSPL folder generated by the function. \strong{CB: we need to go through and document every column in the output of this --i.e. what does INVID and GChar1 etc etc mean -- where is metadata located for this?}

Output NVSPL txt file contains the following N columns (need to go through and doc these):

\itemize{
\item{\strong{SiteID}: Site name.}
\item{\strong{STime}: tbd.}
\item{\strong{H12p5}: tbd}
\item{\strong{all H columns}: tbd.}
\item{\strong{H20000}: tbd.}
\item{\strong{dbA}: ....}
\item{\strong{dbC}: ....}
\item{\strong{dbZ}: ...}
\item{\strong{etc}: ...}
\item{\strong{timezone}: Indicates the timezone reflected in the audio filename. 'GMT' indicates that audio recordings were taken with no timezone adjustment. If recordings were taken in local time at a study site, please specify an \href{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones#List}{Olson-names-formatted character timezone} for the location (e.g., 'America/Los_Angeles'). This is extremely important to foster clarity in data analysis through the years, as some projects vary across time and space in terms of whether recordings are taken in GMT vs. local time.}
}
}
\description{
This function uses PAMGuide code to convert wave files into NVSPL format. PAMGuide was developed by Nathan D. Merchant et al. 2015 (see \strong{Details}). The suggested workflow for this function is to first set test.file = TRUE to test that your workflow has been accurately parameterized. Next, to batch process NVSPLs, run with test.file = FALSE.
}
\details{
This function was developed by the National Park Service Natural Sounds and Night Skies Division to act as a wrapper to PAMGuide that would support NSNSD data processing workflows. PAMGuide was published as supplementary material to the following Open Access journal article:

\href{https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210X.12330}{Merchant, N.D., Fristrup, K.M., Johnson, M.P., Tyack, P.L., Witt, M.J., Blondel, P., Parks, S.E. (2015). Measuring acoustic habitats. Methods in Ecology and Evolution.}

\itemize{
\item{\href{https://besjournals.onlinelibrary.wiley.com/action/downloadSupplement?doi=10.1111\%2F2041-210X.12330&file=mee312330-sup-0001-AppendixS1.pdf}{View accompanying PAMGuide tutorial}}
\item{\href{https://sourceforge.net/projects/pamguide/}{Download original PAMGuide code}}
}
}
\examples{
\dontrun{

# Create an input directory for this example
dir.create('example-input-directory')

# Read in example wave files
data(exampleAudio1)
data(exampleAudio2)

# Write example waves to example input directory
tuneR::writeWave(object = exampleAudio1,
                 filename = 'example-input-directory/Rivendell_20210715_114502.wav')
tuneR::writeWave(object = exampleAudio2,
                 filename = 'example-input-directory/Rivendell_20210715_115502.wav')

# Perform wave_to_nvspl in test mode (test.file = TRUE)
wave_to_nvspl(
 input.directory = 'example-input-directory',
 data.directory = FALSE,
 test.file = TRUE,
 project = 'testproject',
 timezone = 'GMT')

# Perform wave_to_nvspl in batch mode (test.file = FALSE)
wave_to_nvspl(
 input.directory = 'example-input-directory',
 data.directory = FALSE,
 test.file = FALSE,
 project = 'testproject',
 timezone = 'GMT')

# Verify that NVSPL outputs have been created
nvspls <- list.files('example-input-directory/NVSPL', full.names = TRUE)

# View one of the NVSPL outputs
one.nvspl <- read.delim(file = nvspls[1], sep = ',')

# Delete all temporary example files when finished
unlink(x = 'example-input-directory', recursive = TRUE)

}

}
\seealso{
\code{\link{nvspl_to_ai}}
}
